refactor: replace scan/classify definitions with re-exports from common/sources
scan.py and classify.py become backward-compat shims that re-export
SourceFile, FileInfo, scan_sources, classify_sources from the new
scripts.common.sources module. differ.py and run.py import directly
from common. No behaviour change; all 483 tests pass.
